Prepare app structure & Toolbox topbar
Done a comprehensive project setup, adding FontAwesome, Redux, SCSS,
TypeScript and working with React as usual. Defined a basic color
scheme for app, also the folder structure is prepared. Set up a global
state for app and also configured TypeScript for this project.

Last, but not least improvement is creating a template (not fully
decorated) toolbox bar. Now it contains a bunch of buttons, which later
will be used to create and modify shapes made by a user.