Finish designing process for square shape creation
The very first app functionality is ready to use ! A square shape can
be created on top of Workspace grid. Currently, only square size and
position are possible to adjust. All the remain properties are yet
not implemented.

Squares can be created on top of each too, but they cannot be grouped
inside a layout (so far, this feature has not been added).

Modified the toolbox coloring a bit. The main purpose of this change was
to remove transparency from its' background.

Workspace and also the whole app does not allow the created content to
overflow. That means, each time a shape is created, which size goes
beyond the visible screen size, the overflowing part is being cut off
permanently.