Complete shape targeting
This commit includes in-depth update for targeting a shape object:
- Shape is targetable in conditions, where none of action buttons are
chosen, or when either Design or Modify button is currently picked.
- Implement a target escape mechanism, which works either by picking
a new shape object (following above mentioned rules), or when the
same shape object is clicked (which disables any target).
- When any object is targeted, Create type buttons are temporatily
disabled, until no object is picked as a target.
- Modify and Design buttons have special condition for requesting
option list. That is, in order for those buttons to work, a shape
object has to be targeted. There is a checker for that - if a Modify
or Design button is picked, but there is no shape object targeted,
program is waiting for user to pick the shape before displaying
possible options to adjust.

= Implementation part of option for all of Design and Modify buttons
is still undone for a given moment. This functionality will be
available within the next commit.